plant-disease-prediction/server/requirements.txt
fastapi
numpy==1.24.3
pillow==10.3.0
requests
tensorflow
uvicorn
